requirements.txt
rplidar==0.9.2
paho-mqtt==1.5.0
